feat: move observability resolver to general purpose feature_gating_helpers
