--- packages/google-api-core/google/api_core/observability/__init__.py
@@ -1,11 +1,0 @@
-from .options import (
-    clear_test_env_overrides,
-    resolve_feature_flags,
-    set_test_env_override,
-)
-
-__all__ = [
-    "resolve_feature_flags",
-    "set_test_env_override",
-    "clear_test_env_overrides",
-]
